Create and use new hook, fix but re going back
